blart_analysis::datasets

Function load_sods_u64_dataset

pub fn load_sods_u64_dataset(
    path: &Path,
    consume: impl FnMut(&[Unalign<U64>]),
) -> Result<(), Error>

Load an SOSD dataset from the given path and consume the values in-place with the given closure.

The values are unaligned because we’re memory mapping the dataset file to quickly access the keys, and the start pointer won’t necessarily have the right alignment for a u32/u64 value.