blart_analysis

Module key_analysis

This module contains functionality used to analyze key distributions and determine optimal inner node sizes.

Structs

CalculateInnerNodeDist 🔒

This struct contains the running state needed to calculate an InnerNodeWidthDistribution when fed a sorted list of keys.

InnerNodeWidth

This type represents an inner node width in the range [2, 256].

InnerNodeWidthAssignment

This struct represents a set of inner nodes of specific widths and kinds.

InnerNodeWidthDistribution

This struct represents a distribution of inner node widths gathered from a set of keys.

PrefixFrame 🔒

WidthTryFromIntError

This error is returned when attempting to convert from a usize to a InnerNodeWidth, but the usize is out of range.

Enums

InnerNodeKind

This enum represents the different kinds of inner nodes.

Functions

analyze_sorted_key_iter

Given a sorted iterator of keys, compute the distribution of inner node widths.

cost_between 🔒

Calculate the cost for using an inner node of new_width preceded by a inner node of width prev_width with the given kind calculated against the given distribution.