blart_analysis::key_analysis

Function cost_between

fn cost_between(
    prev_width: Option<InnerNodeWidth>,
    new_width: InnerNodeWidth,
    kind: InnerNodeKind,
    dist: &InnerNodeWidthDistribution,
) -> f64

Calculate the cost for using an inner node of new_width preceded by a inner node of width prev_width with the given kind calculated against the given distribution.

This function allows for incrementally calculating the assignment cost by summing the cost between successive pairs of widths in the assignment.