blart_analysis::key_analysis

Struct CalculateInnerNodeDist

struct CalculateInnerNodeDist {
    last_key_bytes: Vec<u8>,
    is_first: bool,
    dist: InnerNodeWidthDistribution,
    prefix_stack: Vec<PrefixFrame>,
}

This struct contains the running state needed to calculate an InnerNodeWidthDistribution when fed a sorted list of keys.

Fields
last_key_bytes: Vec<u8>is_first: booldist: InnerNodeWidthDistributionprefix_stack: Vec<PrefixFrame>
Implementations

impl CalculateInnerNodeDist

fn new() -> Self

fn ingest<K: AsBytes>(&mut self, keys: impl IntoIterator<Item = K>)

Consume the given keys and add the results to the running inner node width distribution.

This function assumes that the list of keys is sorted by their byte representation (as returned by AsBytes::as_bytes), otherwise it will return incorrect results. Also, the keys should be globally sorted over all invocations of ingest.

Panic

This function also assumes that there are no key is a prefix of any other key in the list, other it may panic.

fn finish(self) -> InnerNodeWidthDistribution

Consume the running state to produce the final inner node width distribution.

fn calculate_longest_common_prefix(&mut self, new_key_bytes: &[u8]) -> usize

fn update_last_key_bytes(&mut self, new_key_bytes: &[u8])

fn write_frame_into_dist(dist: &mut InnerNodeWidthDistribution, _: PrefixFrame)