blart_analysis::key_analysis

Struct InnerNodeWidthDistribution

pub struct InnerNodeWidthDistribution {
    counts: [usize; 255],
}

This struct represents a distribution of inner node widths gathered from a set of keys.

Fields
counts: [usize; 255]

This field store an array for each of the possible node width, from [2, 256].

It is not possible to have a node width of 0 or 1 (or greater than 256). Because of that, we store node width 2 at the 0th index, node width 3 at the 1st index, etc etc, and node width 256 at the 254th index.

Implementations

impl InnerNodeWidthDistribution

pub fn add(&mut self, width: InnerNodeWidth, count: usize)

For the InnerNodeWidth increase the corresponding count by count.

pub fn get(&self, width: InnerNodeWidth) -> usize

Retrieve the count for the given InnerNodeWidth

pub fn iter(&self) -> impl Iterator<Item = (InnerNodeWidth, usize)> + use<'_>

Return an iterator over all InnerNodeWidths with non-zero count.

pub fn find_optimal_assignment( &self, num_inner_nodes: usize, ) -> InnerNodeWidthAssignment

Find the optimal assignment of num_inner_nodes - 1 inner nodes using this inner node width distribution as a factor in the cost.

The optimization currently only takes into account the size of the node. See InnerNodeKind for details. A consequence of this is that the optimization results will only ever return inner node assignments of InnerNodeKind::KeyChildCompressed because it is always smaller than InnerNodeKind::ChildOnlyCompressed.

A future version of this function could decide to take into account the relative lookup or update performance of the two kinds and return a different result.